server/prefect-worker/requirements.txt
prefect==3.0.0
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.0
aiosqlite==0.20.0
httpx==0.27.0
python-multipart==0.0.12
openai==1.54.0
pandas==2.2.3
openpyxl==3.1.5
pypdf==5.1.0
mysql-connector-python==9.1.0
python-dotenv==1.0.1

